refactor: move config entry point into frakton-ng/core
- Add config symbols (injection tokens, providers, features, models)
  into core/src/config/ and export from core/public-api.ts
- Update fkt-field.component.ts import from 'frakton-ng/config' to 'frakton-ng/core'
- Update root public-api.ts to re-export from 'frakton-ng/core'
- Remove tsconfig path for 'frakton-ng/config'
- Delete the frakton-ng/config entry point directory
