--- libs/frakton-ng/config/index.ts
@@ -1,5 +1,0 @@
-export * from './injection-tokens';
-export type { FktFieldErrorHandler } from './models/fkt-field-error-handler';
-export { provideFktConfig } from './providers/provide-fkt-config';
-export { withFieldErrorMessages } from './features/with-field-error-messages';
-export { withI18nIntegration } from './features/with-i18n-integration';
