--- libs/frakton-ng/config/ng-package.json
@@ -1,5 +1,0 @@
-{
-  "lib": {
-    "entryFile": "index.ts"
-  }
-}
