+++ libs/frakton-ng/core/public-api.ts
@@ -1,2 +1,9 @@
 export * from './src/fkt-color';
 export * from './src/fkt-identifiable';
+export * from './src/config/injection-tokens';
+export type { FktFieldErrorHandler } from './src/config/fkt-field-error-handler';
+export type { FktTranslationContext } from './src/config/fkt-translation-context';
+export type { RecomputeOn } from './src/config/recompute-on';
+export { provideFktConfig } from './src/config/provide-fkt-config';
+export { withFieldErrorMessages } from './src/config/with-field-error-messages';
+export { withI18nIntegration } from './src/config/with-i18n-integration';
