+++ libs/frakton-ng/core/src/config/fkt-translation-context.ts
