+++ libs/frakton-ng/core/src/config/provide-fkt-config.ts
