+++ libs/frakton-ng/core/src/config/recompute-on.ts
@@ -1,8 +1,6 @@
 import { Signal } from '@angular/core';
 import { Observable } from 'rxjs';
 
-
-
 export type RecomputeOn =
     | Signal<any>
     | Observable<any>
