+++ libs/frakton-ng/public-api.ts
@@ -1,1 +1,1 @@
-export * from './config';
+export * from 'frakton-ng/core';
