+++ tsconfig.json
@@ -16,7 +16,6 @@
             "frakton-ng/card": ["libs/frakton-ng/card"],
             "frakton-ng/checkbox": ["libs/frakton-ng/checkbox"],
             "frakton-ng/color-picker": ["libs/frakton-ng/color-picker"],
-            "frakton-ng/config": ["libs/frakton-ng/config"],
             "frakton-ng/core": ["libs/frakton-ng/core"],
             "frakton-ng/date-picker": ["libs/frakton-ng/date-picker"],
             "frakton-ng/dialog": ["libs/frakton-ng/dialog"],
